Separated phase diagrams and sufrace generation from initial database module. Changed what is exported: Now the InitialDatabase object is exported instead of a dataframe. Added lzma compression to the exported databases.
