+++ src/MatDBForge/core/initial_db.py
@@ -5,9 +5,10 @@ with an applied perturbation with respect to the temperature.
 """
 
 import itertools as it
-import json as js
+import lzma
 import os
 import pathlib
+import pickle
 import re
 import warnings
 from io import BytesIO, TextIOWrapper
@@ -27,233 +28,22 @@ from aiida import orm
 from dscribe.descriptors import SOAP
 from dscribe.kernels import AverageKernel
 from mp_api.client import MPRester
-from pymatgen.core.periodic_table import Element, Species
+from pymatgen.core.periodic_table import Element
 from pymatgen.core.structure import Lattice, Structure
 from pymatgen.core.surface import Slab
 from pymatgen.io.ase import AseAtomsAdaptor
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from slugify import slugify
 
-import MatDBForge.core.structure as mdf_struct
-from MatDBForge.core import exceptions as mdbex
+import MatDBForge as mdb
+import MatDBForge.core.exceptions as mdb_exc
+import MatDBForge.core.phase_diagram as mdb_pd
+import MatDBForge.core.structure as mdb_struct
+import MatDBForge.core.surfaces as mdb_surf
 from MatDBForge.core import utils as ut
 
 # Filtering certain warnings
 warnings.filterwarnings("ignore", category=vasp.outputs.UnconvergedVASPWarning)
-
-
-class BinaryPhaseDiagram:
-    """
-    Class representing a binary phase diagram of materials.
-
-    Parameters:
-        material (str): The name of the material.
-        *phases (Phase): Variable number of Phase objects representing the
-                         phases in the diagram.
-
-    Attributes:
-        phases (list): List of Phase objects representing the phases in the diagram.
-        material (str): The name of the material.
-
-    """
-
-    def __init__(self, material: str, *phases: "Phase"):
-        self.phases = []
-        self.phase_names = []
-        self.material = material
-        self.phase_dict = {}
-
-        for phase in phases:
-            self.add_phase(phase)
-
-        for phase in self.phases:
-            self.phase_names.append(phase.name)
-
-    def add_phase(self, phase):
-        """Add a phase to the phase diagram.
-
-        Parameters:
-            phase (Phase): The Phase object to add.
-
-        """
-        self.phases.append(phase)
-        phase.phasediagram = self.__class__
-        self.phase_dict[phase.name] = phase
-
-    def get_phase(self, phase):
-        """
-        Gets a Phase object from a BinaryPhaseDiagram using either
-        the phase name (as a string) or a Phase object.
-
-        Parameters
-        ----------
-        phase : str | Phase
-            Phase or name of the desired phase.
-
-        Returns
-        -------
-        Phase
-            Phase object corresponding to the given phase
-
-        Raises
-        ------
-        TypeError
-            _description_
-        TypeError
-            _description_
-        """
-
-        if isinstance(phase, Phase):
-            return self.phase_dict[phase.name]
-        if isinstance(phase, str):
-            phase_str = phase
-            phase = self.phase_dict.get(phase_str, None)
-            if phase:
-                return phase
-            else:
-                raise mdbex.PhaseNotFound(self, phase_str)
-        else:
-            raise TypeError("The given phase object is not a Phase-like object.")
-
-    def __repr__(self):
-        """
-        Override attribute lookup.
-        Allows accessing phases by their names as attributes.
-
-        Parameters:
-            name (str): The name of the attribute.
-
-        Returns:
-            Phase: The Phase object with the specified name.
-
-        Raises:
-            AttributeError: If the attribute is not found.
-
-        """
-        repr_str = (
-            f"{self.material} phase diagram with phases:"
-            f" {[phase.name for phase in self.phases]}"
-        )
-
-        return repr_str
-
-    def __getattr__(self, name):
-        if name in self.phase_dict:
-            return self.phase_dict[name]
-        else:
-            raise AttributeError(f"'PhaseDiagram' object has no attribute '{name}'")
-
-
-# TODO: Imlpement this
-class TernaryPhaseDiagram:
-    ...
-
-
-class Phase:
-    """Class representing a phase in a phase diagram.
-
-    Parameters:
-    name: str
-        The name of the phase.
-    base_elem:
-        The base element of the phase.
-    base_elem_comp_max: float
-        The maximum composition of the base element in the phase.
-    base_elem_comp_min: float
-        The minimum composition of the base element in the phase.
-    prototype: str
-        The prototype of the phase.
-    offset: float
-        The offset value of the phase.
-    phase_diagram: PhaseDiagram
-        The parent PhaseDiagram object that the phase belongs to.
-
-    """
-
-    def __init__(
-        self,
-        name: str,
-        base_elem,
-        base_elem_comp_max: float,
-        base_elem_comp_min: float,
-        prototype: str,
-        offset: float = 0,
-        phase_diagram: "BinaryPhaseDiagram" = None,
-    ):
-        self.name = name
-        self.base_elem = Element(base_elem)
-        self.base_elem_comp_max = float(base_elem_comp_max)
-        self.base_elem_comp_min = float(base_elem_comp_min)
-        self.prototype = prototype
-        self.offset = float(offset)
-        self.phase_diagram = phase_diagram
-
-        # if phase_diagram is not None:
-        # phase_diagram.add_phase(self)
-        # self.phase_diagram = phase_diagram.__name__
-
-    def __str__(self):
-        """Return a string representation of the phase.
-
-        Returns:
-            str: The string representation of the phase.
-
-        """
-        repr_string = (
-            f"Phase '{self.name}', {self.base_elem_comp_min*100:.1f}% {self.base_elem} - "
-            f"{self.base_elem_comp_max*100:.1f}% {self.base_elem} (± {self.offset*100:.1f}%)"
-        )
-
-        if self.phase_diagram is not None:
-            # repr_string += f" (belongs to {self.phase_diagram.material})"
-            repr_string += f" (belongs to {self.phase_diagram})"
-
-        return repr_string
-
-    def __key(self):
-        return (
-            self.name,
-            self.base_elem,
-            self.base_elem_comp_max,
-            self.base_elem_comp_min,
-            self.prototype,
-            self.offset,
-            self.phase_diagram,
-        )
-
-    def __eq__(self, other):
-        if not isinstance(other, Phase):
-            # Do not try to compare against different types
-            return NotImplemented
-        return self.__key() == other.__key()
-
-    def __hash__(self):
-        return hash(self.__key())
-
-    def perc_in_phase(self, perc: float, offset: bool = True) -> bool:
-        if perc > 1:
-            perc /= 100
-
-        if offset:
-            offset = self.offset
-        else:
-            offset = 0
-
-        inPhase = (
-            (self.base_elem_comp_min - offset)
-            < perc
-            < (self.base_elem_comp_max + offset)
-        )
-
-        if inPhase:
-            return True
-        else:
-            return False
-
-    def get_base_elem_perc(self, structure) -> float:
-        comp_dict = structure.composition.fractional_composition.as_dict()
-        comp_base = comp_dict.get(self.base_elem.symbol, 0.0)
-        return comp_base
 
 
 class InitialDatabase:
@@ -296,6 +86,8 @@ class InitialDatabase:
     eV2Eh = 1 / Eh2eV
 
     def __init__(self, database_name: str, max_num_atoms: int = 64) -> None:
+        self.db_version = mdb.__version__
+
         # Name of the database
         self.database_name = database_name
 
@@ -313,7 +105,7 @@ class InitialDatabase:
             self.df = self._load_database()
 
         # Loading materials project API key from a json file
-        self.secrets = gather_secrets()
+        self.secrets = ut.gather_secrets()
 
     def __repr__(self):
         # Fields on the dataframe
@@ -342,18 +134,26 @@ class InitialDatabase:
         """
 
         db_path = pathlib.PurePath(self.database_name)
-
-        if len(db_path.suffixes) == 0:
-            suffix = ".pkl"
-        else:
-            suffix = ""
-
-        database = pd.read_pickle(f"{self.database_name}{suffix}")
-        ut.custom_print(f"Loaded '{self.database_name}{suffix}'", "info")
-
         self.database_name = db_path.name.replace(db_path.suffix, "")
 
-        return database
+        if len(db_path.suffixes) == 0:
+            suffix = ".xz"
+        else:
+            suffix = ""
+            if db_path.suffixes[0] == '.pkl':
+                database = pd.read_pickle(f"{self.database_name}{suffix}")
+            elif db_path.suffixes[0] == '.xz':
+                with lzma.open(db_path, 'rb') as f:
+                    database = pickle.load(f)
+
+                    # Setting parameters from InitialDatabase
+                    self.database_name = database.database_name
+                    self.max_num_atoms = database.max_num_atoms
+                    self.db_version = database.db_version
+
+        ut.custom_print(f"Loaded '{self.database_name}{suffix}'", "info")
+
+        return database.df
 
     def _check_database(self) -> bool:
         """
@@ -418,6 +218,7 @@ class InitialDatabase:
             ]
         )
 
+        df.attrs["db_version"] = self.db_version
         ut.custom_print(f"Created database '{self.database_name}'.", "done")
 
         return df
@@ -539,176 +340,11 @@ class InitialDatabase:
         else:
             return False
 
-    # def old_find_repeat_structures(
-    #     self,
-    #     delete=False,
-    #     filters: list = None,
-    # ):
-    #     # Filtering the dataframe
-    #     # Filters allow to select certain subsets of structures
-    #     # from the database.
-    #     filtered_df = self.df
-    #     remaining_df = self.df
-    #     for filt in filters:
-    #         filtered_df = filtered_df.loc[filtered_df[filt]]
-    #         remaining_df = remaining_df.loc[~remaining_df[filt]]
-
-    #     # Getting the unique phases in the dataframe
-    #     phase_list = filtered_df.phase.unique()
-
-    #     # Getting the species from the current phase diagram
-    #     species = CuZnInitialDatabase.ALLOY_SET
-    #     species_str_list = [spec.symbol for spec in species]
-
-    #     # Setting SOAP related parameters
-    #     r_cut = 6
-    #     r_cut = 6
-    #     n_max = 8
-    #     l_max = 6
-
-    #     tot_duplicate_name_list = []
-    #     # Getting the total number of structures
-    #     df_total_struct_list = len(filtered_df.structure.values)
-
-    #     # Computing SOAP for every phase
-    #     for curr_phase in phase_list:
-    #         # Setting up the SOAP descriptor
-    #         soap = SOAP(
-    #             species=species_str_list,
-    #             periodic=True,
-    #             r_cut=r_cut,
-    #             n_max=n_max,
-    #             l_max=l_max,
-    #             # average="inner",
-    #             sparse=False,
-    #         )
-
-    #         soap_structs = []
-
-    #         # Getting the current structures
-    #         structure_list = filtered_df[
-    #             filtered_df.phase == curr_phase
-    #         ].structure.values
-
-    #         # Getting the names fo the current structures
-    #         name_list = filtered_df[
-    #             filtered_df.phase == curr_phase
-    #         ].material_name.values
-
-    #         # Getting the total structure count
-    #         tot_structures = len(structure_list)
-
-    #         # Counter for the total equivalent structure number
-    #         tot_equival = 0
-
-    #         for pym_struct in structure_list:
-    #             ase_struct = AseAtomsAdaptor().get_atoms(pym_struct)
-
-    #             # Create output for multiple systems in parallel
-    #             struct_soap = soap.create(ase_struct, n_jobs=-1, verbose=True)
-    #             print("struct_soap: ", struct_soap)
-
-    #             curr_feat_sum = struct_soap.sum()
-
-    #             soap_structs.append(curr_feat_sum)
-
-    #         soap_arr = np.array(soap_structs)
-    #         dupl_list = []
-    #         for soap in soap_structs:
-    #             curr_soap_arr = np.array(soap)
-    #             comp_arr = np.isclose(curr_soap_arr, soap_arr, rtol=7.5e-04, atol=5e-05)
-
-    #             if np.count_nonzero(comp_arr) > 1:
-    #                 dupl_list.append(soap)
-
-    #             soap_arr = soap_arr[~comp_arr]
-
-    #         tot_equival = len(dupl_list)
-    #         dupl_names = []
-
-    #         dupl_list = set(dupl_list)
-    #         for dup in dupl_list:
-    #             soap_1_name = name_list[soap_structs.index(dup)]
-    #             dupl_names.append(soap_1_name)
-    #             tot_duplicate_name_list.append(soap_1_name)
-
-    #         ut.custom_print(
-    #             f"Total selected structures: {tot_structures}, equivalent: {tot_equival} "
-    #             f"({(tot_equival/tot_structures)*100:.2f}%)",
-    #             "debug",
-    #         )
-
-    #     duplicate_structures = tot_duplicate_name_list
-    #     print("duplicate_structures: ", duplicate_structures)
-
-    #     if delete:
-    #         ut.custom_print(
-    #             f"{len(duplicate_structures)} structures marked for deletion.",
-    #             "debug",
-    #         )
-
-    #         # base_struct_names = list(
-    #         #     set(self.df.loc[self.df.base].material_name.values)
-    #         # )
-    #         # duplicate_structures.extend(base_struct_names)
-
-    #         print("filtered_df: ", filtered_df.shape)
-    #         # duplicate_structures_df = filtered_df[~filtered_df.material_name.isin(duplicate_structures)]
-
-    #         # TODO: Check if this works and then drop the matching idx.
-    #         for idx, row in filtered_df.iterrows():
-    #             # print('idx: ', idx)
-    #             for name in duplicate_structures:
-    #                 if name == row.material_name:
-    #                     break
-    #             else:
-    #                 # print(name)
-    #                 pass
-
-    #         duplicate_structures_df = filtered_df.loc[
-    #             ~filtered_df["material_name"].isin(duplicate_structures)
-    #         ]
-    #         print("duplicate_structures_df: ", duplicate_structures_df["material_name"])
-    #         print("unique_structures_df: ", duplicate_structures_df.shape)
-    #         # quit()
-
-    #         # unique_structures_df_drop = unique_structures_df.drop_duplicates(
-    #         #     subset=["material_name"],
-    #         #     keep="first",
-    #         #     # inplace=True,
-    #         # )
-    #         # print('unique_structures_df_drop: ', unique_structures_df_drop.shape)
-
-    #         unique_structures_df_drop = pd.concat(
-    #             (remaining_df, duplicate_structures_df)
-    #         )
-    #         print("unique_structures_df_drop concat: ", unique_structures_df_drop.shape)
-
-    #         # for idx, row in unique_structures_df_drop.iterrows():
-    #         #     poscar = vasp.Poscar(row.structure)
-    #         #     poscar.write_file(f'/tmp/export_poscars_test/{row.material_name}.poscar')
-    #         # quit()
-
-    #         # print("unique_structures_df: ", unique_structures_df)
-
-    #         self.df = unique_structures_df_drop
-    #         ut.custom_print(
-    #             f"Deleted {len(duplicate_structures)} structures.",
-    #             "warn",
-    #         )
-
-    #     else:
-    #         ut.custom_print(
-    #             f"{len(duplicate_structures)} repeated structures found. "
-    #             "Database untouched as 'delete' is set to False.",
-    #             "info",
-    #         )
-
     def find_repeat_structures(
         self,
         delete=False,
         filters: list = None,
-        phase: Phase = None,
+        phase: mdb_pd.Phase = None,
     ):
         # Filtering the dataframe
         # Filters allow to select certain subsets of structures
@@ -723,7 +359,9 @@ class InitialDatabase:
                         (filtered_df[filt[0]]) & (filtered_df[filt[1]])
                     ]
 
-                    remaining_df = remaining_df.loc[~((remaining_df[filt[0]]) & (remaining_df[filt[1]]))]
+                    remaining_df = remaining_df.loc[
+                        ~((remaining_df[filt[0]]) & (remaining_df[filt[1]]))
+                    ]
 
                 else:
                     filtered_df = filtered_df.loc[filtered_df[filt]]
@@ -735,13 +373,13 @@ class InitialDatabase:
             if isinstance(phase, list):
                 for curr_phase in phase:
                     if isinstance(curr_phase, "str"):
-                        curr_phase = self.CUZN_PHASES.get_phase(phase)
+                        curr_phase = self.DB_PHASE_DIAGRAM.get_phase(phase)
 
                     phase_list.append(curr_phase)
 
             else:
                 if isinstance(phase, str):
-                    phase = self.CUZN_PHASES.get_phase(phase)
+                    phase = self.DB_PHASE_DIAGRAM.get_phase(phase)
                 phase_list = [phase]
 
         # If no phase is given, getting the unique phases in the dataframe
@@ -792,7 +430,7 @@ class InitialDatabase:
             tot_equival = 0
 
             for pym_struct in structure_list:
-                # Converting to ase stuctuer
+                # Converting to ase structure
                 ase_struct = AseAtomsAdaptor().get_atoms(pym_struct)
 
                 # Create soap descriptors for current system and storing it
@@ -865,13 +503,13 @@ class InitialDatabase:
         with MPRester(self.secrets["API_KEY"]) as mpr:
             query_result = mpr.summary.search(material_ids=missing_mat)
             for material in query_result:
-                for phase in CuZnInitialDatabase.CUZN_PHASES.phases:
+                for phase in CuZnInitialDatabase.DB_PHASE_DIAGRAM.phases:
                     if phase.prototype == material.material_id:
                         curr_phase = phase.name
                     else:
                         curr_phase = np.nan
 
-                curr_struct = mdf_struct.Bulk(
+                curr_struct = mdb_struct.Bulk(
                     material_id=str(material.material_id),
                     structure=material.structure,
                     temperature=np.nan,
@@ -903,7 +541,7 @@ class InitialDatabase:
         if target_structures:
             selection_criteria = target_structures
         else:
-            selection_criteria = CuZnInitialDatabase.CUZN_PHASES.keys()
+            selection_criteria = CuZnInitialDatabase.DB_PHASE_DIAGRAM.keys()
 
         folders = read_path.glob("./*")
         list_dir = [
@@ -924,16 +562,18 @@ class InitialDatabase:
             curr_run = vasp.Vasprun(xml_path, parse_potcar_file=False)
 
             # Gathering phase information
-            for phase in CuZnInitialDatabase.CUZN_PHASES.phase_names:
+            for phase in CuZnInitialDatabase.DB_PHASE_DIAGRAM.phase_names:
                 for folder in xml_path.parts:
                     if slugify(folder) == slugify(phase):
                         curr_phase = phase
-                        curr_phase = CuZnInitialDatabase.CUZN_PHASES.get_phase(phase)
+                        curr_phase = CuZnInitialDatabase.DB_PHASE_DIAGRAM.get_phase(
+                            phase
+                        )
                         curr_mat_id = curr_phase.prototype
                         curr_name = f"base_relax_{curr_phase.name}_MP"
 
             # Creating the structure object
-            curr_struct = mdf_struct.Structure().from_vasprun(
+            curr_struct = mdb_struct.Structure().from_vasprun(
                 vasprun=curr_run,
                 base=True,
                 phase=curr_phase,
@@ -963,16 +603,20 @@ class InitialDatabase:
                 String that will be added at the beginning of the filename.
         """
         if suffix:
-            filename = self.database_name + f"_{suffix}.pkl"
+            filename = self.database_name + f"_{suffix}.xz"
         else:
-            filename = self.database_name + ".pkl"
+            filename = self.database_name + ".xz"
 
         if not path:
             path = ""
 
-        file_path = pathlib.Path(path, filename)
 
-        self.df.to_pickle(path=file_path)
+        file_path = pathlib.Path(path, filename)
+        # self.df.to_pickle(path=file_path, compression="gzip")
+
+        with lzma.open(file_path, "wb") as f:
+            pickle.dump(self, f)
+
         ut.custom_print(f"Database saved in {file_path}", "info")
 
     def _apply_user_filters(self, filters: list, target_entries: pd.DataFrame):
@@ -1028,7 +672,7 @@ class InitialDatabase:
                 mat_str = f"{entry.material_name}_perturb_gauss_{perturb_repeat_idx+1}"
 
                 # Creating a new Structure from the perturbed structure structure
-                curr_struct = mdf_struct.Structure(
+                curr_struct = mdb_struct.Structure(
                     material_name=mat_str,
                     structure=new_struct_perturb,
                     material_id=str_matid,
@@ -1045,9 +689,9 @@ class InitialDatabase:
 
                 # Converting the structure to the appropiate type
                 if entry.bulk:
-                    curr_struct_conv = mdf_struct.Bulk().from_mdb_structure(curr_struct)
+                    curr_struct_conv = mdb_struct.Bulk().from_mdb_structure(curr_struct)
                 elif entry.surface:
-                    curr_struct_conv = mdf_struct.Surface().from_mdb_structure(
+                    curr_struct_conv = mdb_struct.Surface().from_mdb_structure(
                         curr_struct,
                         entry.surface_miller,
                     )
@@ -1062,7 +706,6 @@ class InitialDatabase:
 
     def _apply_gauss_perturb(self, structure: Structure, center: float = 0.04):
         new_structure = structure.copy()
-
         new_structure.perturb(distance=0.08, min_distance=0.02)
         return new_structure
 
@@ -1092,7 +735,7 @@ class InitialDatabase:
                 )
 
                 # Creating a new Structure from the perturbed structure structure
-                curr_struct = mdf_struct.Structure(
+                curr_struct = mdb_struct.Structure(
                     material_name=mat_str,
                     structure=new_struct_perturb,
                     material_id=str_matid,
@@ -1109,9 +752,9 @@ class InitialDatabase:
 
                 # Converting the structure to the appropiate type
                 if entry.bulk:
-                    curr_struct_conv = mdf_struct.Bulk().from_mdb_structure(curr_struct)
+                    curr_struct_conv = mdb_struct.Bulk().from_mdb_structure(curr_struct)
                 elif entry.surface:
-                    curr_struct_conv = mdf_struct.Surface().from_mdb_structure(
+                    curr_struct_conv = mdb_struct.Surface().from_mdb_structure(
                         curr_struct
                     )
                 else:
@@ -1183,7 +826,7 @@ class CuZnInitialDatabase(InitialDatabase):
     """
 
     # CuZn alloy phase diagram data
-    alpha_phase = Phase(
+    alpha_phase = mdb_pd.Phase(
         name="alpha",
         base_elem="Zn",
         base_elem_comp_min=0,
@@ -1191,7 +834,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-30",
         offset=0.03,
     )
-    m1 = Phase(
+    m1 = mdb_pd.Phase(
         name="m1",
         base_elem="Zn",
         base_elem_comp_min=0.3895,
@@ -1199,7 +842,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-30",
         offset=0.03,
     )
-    beta_prime = Phase(
+    beta_prime = mdb_pd.Phase(
         name="beta-prime",
         base_elem="Zn",
         base_elem_comp_min=0.455,
@@ -1207,7 +850,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-987",
         offset=0.05,
     )
-    m2 = Phase(
+    m2 = mdb_pd.Phase(
         name="m2",
         base_elem="Zn",
         base_elem_comp_min=0.51,
@@ -1215,7 +858,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-987",
         offset=0.05,
     )
-    gamma = Phase(
+    gamma = mdb_pd.Phase(
         name="gamma",
         base_elem="Zn",
         base_elem_comp_min=0.577,
@@ -1223,7 +866,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-1368",
         offset=0.03,
     )
-    m3 = Phase(
+    m3 = mdb_pd.Phase(
         name="m3",
         base_elem="Zn",
         base_elem_comp_min=0.706,
@@ -1231,7 +874,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-1216020",
         offset=0.05,
     )
-    delta = Phase(
+    delta = mdb_pd.Phase(
         name="delta",
         base_elem="Zn",
         base_elem_comp_min=0.7302,
@@ -1239,7 +882,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-1215518",
         offset=0.05,
     )
-    epsilon = Phase(
+    epsilon = mdb_pd.Phase(
         name="epsilon",
         base_elem="Zn",
         base_elem_comp_min=0.785,
@@ -1247,7 +890,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-972042",
         offset=0.05,
     )
-    m4 = Phase(
+    m4 = mdb_pd.Phase(
         name="m4",
         base_elem="Zn",
         base_elem_comp_min=0.883,
@@ -1255,7 +898,7 @@ class CuZnInitialDatabase(InitialDatabase):
         prototype="mp-79",
         offset=0.01,
     )
-    eta = Phase(
+    eta = mdb_pd.Phase(
         name="eta",
         base_elem="Zn",
         base_elem_comp_min=0.9725,
@@ -1264,7 +907,7 @@ class CuZnInitialDatabase(InitialDatabase):
         offset=0.01,
     )
 
-    CUZN_PHASES = BinaryPhaseDiagram(
+    DB_PHASE_DIAGRAM = mdb_pd.BinaryPhaseDiagram(
         "CuZn", alpha_phase, m1, beta_prime, m2, gamma, m3, delta, epsilon, m4, eta
     )
 
@@ -1293,7 +936,7 @@ class CuZnInitialDatabase(InitialDatabase):
 
     def _get_phase_from_id(self, idx: str) -> str:
         """
-        Searches for the corresponding phase in the self.CUZN_PHASES
+        Searches for the corresponding phase in the self.DB_PHASE_DIAGRAM
         dict to a given a material projects id.
 
         Parameters
@@ -1309,7 +952,9 @@ class CuZnInitialDatabase(InitialDatabase):
 
         # Creating a list of the phase diagram phase names
         phase_list = [
-            phase.name for phase in self.CUZN_PHASES.phases if phase.prototype == idx
+            phase.name
+            for phase in self.DB_PHASE_DIAGRAM.phases
+            if phase.prototype == idx
         ]
 
         # Returning the key if found
@@ -1345,8 +990,8 @@ class CuZnInitialDatabase(InitialDatabase):
             for ind in range(len(structure.species)):
                 new_structure.replace(
                     ind,
-                    # Species(self.CUZN_PHASES.get_phase(phase).base_elem),
-                    Element(self.CUZN_PHASES.get_phase(phase).base_elem),
+                    # Species(self.DB_PHASE_DIAGRAM.get_phase(phase).base_elem),
+                    Element(self.DB_PHASE_DIAGRAM.get_phase(phase).base_elem),
                 )
 
             # Returning new structure with atoms replaced with Cu
@@ -1390,15 +1035,15 @@ class CuZnInitialDatabase(InitialDatabase):
         ------
         KeyError
             Raised if the given phase is not found. All of the available phases
-            are given on the self.CUZN_PHASES dictionary. More phases could be added
+            are given on the self.DB_PHASE_DIAGRAM dictionary. More phases could be added
             there if necessary.
         """
         # Checking for correct phase input
         phase_name = slugify(phase.name)
-        if not self.CUZN_PHASES.get_phase(phase_name):
+        if not self.DB_PHASE_DIAGRAM.get_phase(phase_name):
             raise KeyError(
                 "Wrong phase given. "
-                f"Please introduce one of: {[k for k in self.CUZN_PHASES.phases]}"
+                f"Please introduce one of: {[k for k in self.DB_PHASE_DIAGRAM.phases]}"
             )
 
         # Reading structure from database
@@ -1409,7 +1054,7 @@ class CuZnInitialDatabase(InitialDatabase):
                 material_id_prefix = query_result.material_id.values[0]
                 structure = query_result.structure.values[0]
             except IndexError:
-                raise mdbex.BaseStructureNotFound()
+                raise mdb_exc.BaseStructureNotFound()
 
         # Querying CuZn alpha prototype structure
         else:
@@ -1442,7 +1087,7 @@ class CuZnInitialDatabase(InitialDatabase):
             idxs_str = "".join(map(str, idxs))
 
             # Creating a new bulk from the supercell
-            curr_bulk = mdf_struct.Bulk(
+            curr_bulk = mdb_struct.Bulk(
                 material_name=f"{material_id_prefix}_{phase.name}_super-{idxs_str}",
                 material_id=material_id_prefix,
                 structure=structure,
@@ -1466,13 +1111,13 @@ class CuZnInitialDatabase(InitialDatabase):
     def _create_symmetrical_prototype(
         self,
         structure: Structure,
-        phase: Phase,
+        phase: mdb_pd.Phase,
         # query_result: emmet.core.summary.SummaryDoc,
         # read: bool,
-        structure_obj: mdf_struct.Structure,
+        structure_obj: mdb_struct.Structure,
     ):
         phase = structure_obj.phase
-        curr_phase_atom = self.CUZN_PHASES.get_phase(phase).base_elem
+        curr_phase_atom = self.DB_PHASE_DIAGRAM.get_phase(phase).base_elem
         base_atom_set = list(self.ALLOY_SET - {curr_phase_atom})
 
         new_structure = structure.copy(sanitize=True)
@@ -1497,7 +1142,7 @@ class CuZnInitialDatabase(InitialDatabase):
         material_id_prefix = structure_obj.material_id
 
         # Generating the symmetrized structure
-        new_struct_symm = mdf_struct.Structure(
+        new_struct_symm = mdb_struct.Structure(
             material_name=f"{material_id_prefix}_{phase.name}_symm",
             material_id=material_id_prefix,
             structure=structure,
@@ -1512,7 +1157,7 @@ class CuZnInitialDatabase(InitialDatabase):
         )
 
         if structure_obj.bulk:
-            final_struct = mdf_struct.Bulk().from_mdb_structure(
+            final_struct = mdb_struct.Bulk().from_mdb_structure(
                 mdb_structure=new_struct_symm,
                 new_structure=structure,
             )
@@ -1553,59 +1198,6 @@ class CuZnInitialDatabase(InitialDatabase):
             ) + max_base_elem
 
         return subst_base_elem_perc
-
-    def _gen_perc_surfaces(
-        self,
-        phase: Phase,
-        num_struct: int,
-        current_perc: float,
-        relative=True,
-    ) -> list:
-        """
-        Generate num_struct percentages for a structure in a given phase.
-        The percentages represent the ratio of the base element of the
-        structure's phase.
-
-        Parameters
-        ----------
-        phase : Phase
-            Phase that will be used to define the limits of the percentage values by
-            checking its base_elem range.
-        num_struct : int
-            Number of percentages that will be generated
-        current_perc : float
-            Percentage of the current structure
-        relative : bool, optional
-            Whether to return the percentages by themselves or relative
-            to the percentage of the current structure, by default True
-
-        Returns
-        -------
-        list[float]
-            List of floats containing the generated percentages.
-        """
-
-        # Getting offset. If not found set to 0.
-        offset = phase.offset
-
-        # Randomly generating base_elem percentages for the new structures
-        max_base_elem = (phase.base_elem_comp_min) + offset
-        if max_base_elem > 1:
-            max_base_elem = 1
-
-        min_base_elem = (phase.base_elem_comp_max) - offset
-        if min_base_elem < 0:
-            min_base_elem = 0
-
-        subst_base_elem_perc = (min_base_elem - max_base_elem) * np.random.ranf(
-            size=num_struct
-        ) + max_base_elem
-
-        if relative:
-            adjusted_perc = [(per - current_perc) for per in subst_base_elem_perc]
-            return adjusted_perc
-        else:
-            return subst_base_elem_perc
 
     def _fit_replacements_phase(
         self,
@@ -1655,7 +1247,7 @@ class CuZnInitialDatabase(InitialDatabase):
             rng = np.random.default_rng()
 
         if isinstance(
-            structure, (mdf_struct.Structure, mdf_struct.Surface, mdf_struct.Bulk)
+            structure, (mdb_struct.Structure, mdb_struct.Surface, mdb_struct.Bulk)
         ):
             # structure_obj = structure
             structure = structure.structure
@@ -1717,10 +1309,21 @@ class CuZnInitialDatabase(InitialDatabase):
 
         # Creating a new pymatgen structure using the base one as a template
         new_structure = structure.copy(sanitize=True)
+        site_props_before = structure.site_properties
 
         # Replacing atoms in the structures
         for ind in other_elem_choices:
             new_structure.replace(ind, other_elem)
+
+        # TODO: Instead of this, create a new structure
+        # Copying site properties
+        new_structure = new_structure.copy(
+            sanitize=True, site_properties=site_props_before
+        )
+
+        # if None in new_structure.site_properties['selective_dynamics']:
+        #     print('none found')
+        #     print(new_structure)
 
         return new_structure
 
@@ -1762,7 +1365,7 @@ class CuZnInitialDatabase(InitialDatabase):
         ------
         KeyError
             Raised if the given phase is not found. All of the available phases
-            are given on the self.CUZN_PHASES dictionary. More phases could be added
+            are given on the self.DB_PHASE_DIAGRAM dictionary. More phases could be added
             there if necessary.
         """
         # Instantiating RNG
@@ -1793,7 +1396,6 @@ class CuZnInitialDatabase(InitialDatabase):
             structure = self._create_symmetrical_prototype(
                 structure=structure, phase=phase, structure_obj=structure_obj
             )
-
             # Preparing an array of randomly generated base elem percentages
             # for the new structures
             subst_base_elem_perc = self._gen_base_elem_perc(phase, num_struct)
@@ -1810,7 +1412,6 @@ class CuZnInitialDatabase(InitialDatabase):
             n_at_replacement_upd = self._fit_replacements_phase(
                 phase, structure, subst_base_elem_perc
             )
-
             # Replacing the atoms and generate 'num_replacements'
             # structures for each percentage
             for str_ind, n_atoms in enumerate(n_at_replacement_upd):
@@ -1826,7 +1427,7 @@ class CuZnInitialDatabase(InitialDatabase):
                     )
 
                     # Creating a new Bulk object for the structure with replacement
-                    new_struct_symm = mdf_struct.Bulk(
+                    new_struct_symm = mdb_struct.Bulk(
                         material_name=f"{prototype}_{phase.name}_super-{supercell_vec_str}-{supr_idx}_replacement-{str_ind+1}-{repl+1}",
                         material_id=prototype,
                         structure=new_structure,
@@ -1883,359 +1484,6 @@ class CuZnInitialDatabase(InitialDatabase):
 
         return curr_miller
 
-    def _slab_to_bottom(
-        self,
-        slab: Union[Slab, Structure],
-        offset: int = 2,
-    ) -> Structure:
-        """
-        Move the slab towards the bottom of the cell, leaving a
-        offset wide margin at the bottom.
-
-        Parameters
-        ----------
-        slab : Union[Slab, Structure]
-            Target slab to move to the bottom.
-        miller: tuple
-            Miller index of the slab
-        offset : int, optional
-            Separation to be left between the bottom of the cell
-            and the slab, by default 2, in Angstrom.
-
-
-        Returns
-        -------
-        Structure
-            Pymatgen structure containing slab placed on the bottom,
-            with the same attributes as the original.
-        """
-
-        # Getting the position closest to the bottom
-        bottom = min(slab.cart_coords[:, 2])
-        bottom_arr = np.zeros(shape=slab.cart_coords.shape)
-
-        # Applying the offset
-        bottom_arr[:, 2] += bottom - offset
-
-        # Substracting the bottom position from the slab plus an offset
-        modified_coords = slab.cart_coords - bottom_arr
-
-        new_slab = Structure(
-            lattice=slab.lattice,
-            species=slab.species,
-            coords=modified_coords,
-            coords_are_cartesian=True,
-        )
-
-        return new_slab
-
-    def _check_correct_vacuum_size(
-        self,
-        slab: Union[Slab, Structure],
-        vacuum_size: float,
-        tolerance: float = 0.5,
-    ) -> bool:
-        # Getting 'c' size for the cell
-        vec_c = slab.lattice.c
-
-        # Getting position of the topmost layer
-        z_axis_max = max(slab.cart_coords[:, 2])
-
-        # Getting vacuum layer thickness by substracting
-        vac_layer_thickness = vec_c - z_axis_max
-
-        # Checking if layer is greater or equal than vacuum_size
-        if abs(vac_layer_thickness - vacuum_size) <= tolerance:
-            return True
-        else:
-            return False
-
-    def _adjust_vacuum(self, slab: Slab, vacuum_size: float) -> Slab:
-        # Getting 'c' vector for the cell
-        vec_c = slab.lattice.c
-
-        # Getting position of the topmost layer
-        z_axis_max = max(slab.cart_coords[:, 2])
-
-        #
-        current_vacuum_size = vec_c - z_axis_max
-
-        # Computing correct slab size
-        corr_slab_size = z_axis_max + vacuum_size
-
-        # Computing the difference between the correct slab
-        diff = vec_c - corr_slab_size
-        # print('vec_c: ', vec_c)
-        # print('top layer: ', z_axis_max)
-        # print('calculated distance:',vec_c-z_axis_max)
-        # print('corr_slab_size: ', corr_slab_size)
-
-        # Changing the 'c' vector
-        if current_vacuum_size > vacuum_size:
-            new_vec_c = vec_c - diff
-        elif current_vacuum_size < vacuum_size:
-            new_vec_c = vec_c + diff
-
-        # Creating a new abc vector
-        new_latt_abc = np.array(slab.lattice.abc)
-        new_latt_abc[-1] = new_vec_c
-
-        # Converting the abc vector into a 3x3 matrix
-        new_latt_matrix = np.zeros([3, 3])
-        diag = np.diag_indices(3)
-        new_latt_matrix[diag] = new_latt_abc
-
-        # Creating a lattice use the new matrix
-        new_lattice = Lattice(matrix=new_latt_matrix)
-
-        # Using the lattice to create a new slab
-        new_slab = Structure(
-            lattice=new_lattice,
-            species=slab.species,
-            coords=slab.cart_coords,
-            coords_are_cartesian=True,
-        )
-        return new_slab
-
-    def _make_clean_surf(
-        self,
-        bulk: Union[Structure, Slab],
-        max_num_at: float,
-        n_layers: int,
-        miller_list: list,
-        fixed: int,
-    ):
-        img_miller = []
-        images = []
-
-        for miller in miller_list:
-            # Object that allows to generate the slab.
-            # The attach_grap parameter is disabled to increase speed.
-            gen = cts.SlabGenerator(
-                bulk,
-                miller_index=(miller),
-                layers=n_layers,
-                attach_graph=False,
-                layer_type="angs",
-                fixed=fixed,
-                standardize_bulk="True",
-                vacuum=7.5,
-            )
-
-            # Getting unique terminations for the current surface
-            termination = gen.get_unique_terminations()
-
-            for ind, t in enumerate(termination):
-                img_miller.append(miller)
-                imgsize = gen.get_slab(iterm=ind).get_global_number_of_atoms()
-                slab_rep = int(max_num_at / imgsize)
-
-                try:
-                    slab = gen.get_slab(iterm=ind, size=slab_rep)
-                except Exception:
-                    break
-
-                images.append(slab)
-
-        return images, img_miller
-
-    def _gen_slab_pool(self, miller, bulk, max_num_at, n_layers, fixed):
-        img_miller = []
-        images = []
-
-        gen = cts.SlabGenerator(
-            bulk,
-            miller_index=(miller),
-            layers=n_layers,
-            layer_type="angs",
-            fixed=fixed,
-            standardize_bulk="True",
-            vacuum=7.5,
-        )
-
-        # Getting unique terminations for the current
-        # surface
-        termination = gen.get_unique_terminations()
-
-        for ind, t in enumerate(termination):
-            img_miller.append(miller)
-            imgsize = gen.get_slab(iterm=ind).get_global_number_of_atoms()
-            slab_rep = int(max_num_at / imgsize)
-
-            try:
-                slab = gen.get_slab(iterm=ind, size=slab_rep)
-            except Exception:
-                break
-
-            images.append(slab)
-
-        return images, img_miller
-
-    def _make_clean_surf_mp(
-        self,
-        bulk: Union[Structure, Slab],
-        max_num_at: float,
-        n_layers: int,
-        miller_list: list,
-        fixed: int,
-    ):
-        img_miller = []
-        images = []
-
-        # gen = make_generator_slab
-        # Original parameters:
-        # fixed = 3
-
-        with Pool() as p:
-            slabs_worker = p.starmap(
-                self._gen_slab_pool,
-                zip(
-                    miller_list,
-                    it.repeat(bulk),
-                    it.repeat(max_num_at),
-                    it.repeat(n_layers),
-                    it.repeat(fixed),
-                ),
-            )
-            for slb, mill in slabs_worker:
-                if isinstance(slb, list):
-                    for i in slb:
-                        images.append(i)
-                if isinstance(mill, list):
-                    for m in mill:
-                        img_miller.append(m)
-
-        return images, img_miller
-
-    def __gen_curr_surface(
-        self,
-        phase,
-        curr_bulk_ase,
-        n_layers,
-        n_at,
-        max_miller_index,
-        fixed_layers,
-        get_supercells,
-        limit_per_phase,
-    ):
-        # Filtering specific catkit warnings
-        warnings.filterwarnings("ignore", category=UserWarning)
-        warnings.filterwarnings("ignore", category=RuntimeWarning)
-        n_layers = int(n_layers)
-        n_at = int(n_at)
-
-        slabs = []
-        miller = cts.get_unique_indices(
-            bulk=curr_bulk_ase,
-            max_index=max_miller_index,
-        )
-
-        slabs, miller_idx_slabs = self._make_clean_surf(
-            bulk=curr_bulk_ase,
-            n_layers=n_layers,
-            max_num_at=n_at,
-            miller_list=miller,
-            fixed=fixed_layers,
-        )
-
-        # Will contain tuples as such: (Structure, miller_index_string)
-        slabs_bottom = []
-        for ind, (slab, mill) in enumerate(zip(slabs, miller_idx_slabs)):
-            curr_surf_pymg = AseAtomsAdaptor().get_structure(slab)
-            slab = self._slab_to_bottom(curr_surf_pymg)
-            mill_str = self._get_miller_index_str(mill)
-
-            # INFO: The _adjust_vacuum function does not work correctly.
-            # As of now, the vacuum size is being defined during slab creation,
-            # I suspect it is related to the way pymatgen handles lattices.
-            #
-            # if not self._check_correct_vacuum_size(slab, min_vacuum_size):
-            #     slab = self._adjust_vacuum(slab, min_vacuum_size)
-
-            slabs_bottom.append((slab, mill_str))
-
-        prototype = phase.prototype
-        generated_structures = []
-
-        # Getting only the slabs and their miller index whose total size
-        # is smaller than the maximum given for the InitialDatabase.
-        slabs_size = [
-            (slab, mill)
-            for slab, mill in slabs_bottom
-            if len(slab.sites) < self.max_num_atoms
-        ]
-
-        # Storing the remaining slabs.
-        for idx, (slab, mill) in enumerate(slabs_size):
-            # Getting the current slab's miller index
-            mill_str = self._get_miller_index_str(mill)
-
-            # Preparing the structure name
-            surf_name = (
-                f"{prototype}_{phase.name}_pure_surface_{mill_str}-{idx+1}"
-                f"_{n_layers}-layers_{n_at}-max-at"
-            )
-
-            # Creating a new surface from the supercell
-            curr_strct = mdf_struct.Surface(
-                material_name=surf_name,
-                material_id=prototype,
-                surface_miller=mill_str,
-                structure=slab,
-                temperature=np.nan,
-                perturb=False,
-                base=False,
-                calc_performed=False,
-                phase=phase,
-            )
-
-            # Saving the bulk to the db.
-            self.df = curr_strct.save_to_db(self.df)
-            generated_structures.append(curr_strct)
-
-        # Getting supercells
-        if get_supercells:
-            for idx, (slab, mill) in enumerate(slabs_size):
-                super_list, idx_list, supercells = self._find_supercell_indices(
-                    structure=curr_surf_pymg,
-                    max_atoms=self.max_num_atoms,
-                    get_different_supercells=True,
-                    initial_supercell_size=3,
-                    verbose=False,
-                )
-
-                # Storing the supercells.
-                for supercell, idx, sup_vec in zip(super_list, idx_list, supercells):
-                    if len(supercell.sites) <= self.max_num_atoms:
-                        # Dragging the slab to the bottom
-                        supercell_bottom = self._slab_to_bottom(curr_surf_pymg)
-
-                        # Preparing the structure name
-                        surf_name = (
-                            f"{prototype}_{phase.name}_pure_surface-"
-                            f"{n_layers}-layers_{n_at}-max-at_{self._get_miller_index_str(mill)}-super-{idx+1}"
-                        )
-
-                        # Creating a new surface from the supercell
-                        curr_strct = mdf_struct.Surface(
-                            material_name=surf_name,
-                            material_id=prototype,
-                            structure=supercell_bottom,
-                            temperature=np.nan,
-                            perturb=False,
-                            base=False,
-                            calc_performed=False,
-                            phase=phase,
-                            supercell=sup_vec,
-                        )
-
-                        # Saving the bulk to the db.
-                        self.df = curr_strct.save_to_db(self.df)
-                        generated_structures.append(curr_strct)
-
-        return generated_structures
-        # return surf_name
-
     def _get_structs_current_phase(self, phase):
         # Getting all of the base structures
         base_structs = self.df.loc[self.df.base]
@@ -2249,7 +1497,7 @@ class CuZnInitialDatabase(InitialDatabase):
 
     def generate_surfaces_pure(
         self,
-        phase: Phase,
+        phase: mdb_pd.Phase,
         num_diff_layer_size: int,
         max_miller_index: int = 2,
         min_slab_size: float = 3,
@@ -2314,7 +1562,7 @@ class CuZnInitialDatabase(InitialDatabase):
         """
         # Getting the current phase from the phase name.
         if isinstance(phase, str):
-            phase = CuZnInitialDatabase.CUZN_PHASES.get_phase(phase)
+            phase = CuZnInitialDatabase.DB_PHASE_DIAGRAM.get_phase(phase)
 
         base_structs = self._get_structs_current_phase(phase)
 
@@ -2327,7 +1575,7 @@ class CuZnInitialDatabase(InitialDatabase):
                 "running this function."
             )
 
-            raise mdbex.BaseStructureNotFound(err_msg)
+            raise mdb_exc.BaseStructureNotFound(err_msg)
 
         # Generating an initial random slab size close to the minimum slab thickness
         # given.
@@ -2360,12 +1608,12 @@ class CuZnInitialDatabase(InitialDatabase):
                     self.max_num_atoms,
                     int(self.max_num_atoms / curr_surf_nat),
                 )
-            print("self.max_num_atoms: ", self.max_num_atoms)
-            print("curr_surf_nat: ", curr_surf_nat)
-            print(
-                "int(self.max_num_atoms/curr_surf_nat): ",
-                int(self.max_num_atoms / curr_surf_nat),
-            )
+            # print("self.max_num_atoms: ", self.max_num_atoms)
+            # print("curr_surf_nat: ", curr_surf_nat)
+            # print(
+            #     "int(self.max_num_atoms/curr_surf_nat): ",
+            #     int(self.max_num_atoms / curr_surf_nat),
+            # )
 
             # Getting an ASE Atoms object
             curr_bulk_ase = AseAtomsAdaptor().get_atoms(curr_bulk)
@@ -2420,12 +1668,13 @@ class CuZnInitialDatabase(InitialDatabase):
             with live:
                 while not overall_progress.finished:
                     for n_layers, n_at in total_slabs_gen:
-                        sub_task = job_progress.add_task(
-                            description=f"{int(n_layers)} layers, {int(n_at)} atoms:",
-                            total=None,
-                        )
+                        # sub_task = job_progress.add_task(
+                        #     description=f"{int(n_layers)} layers, {int(n_at)} atoms:",
+                        #     total=None,
+                        # )
 
-                        gen_slabs = self.__gen_curr_surface(
+                        gen_slabs = mdb_surf._gen_curr_surface(
+                            db_obj=self,
                             phase=phase,
                             curr_bulk_ase=curr_bulk_ase,
                             n_layers=n_layers,
@@ -2438,8 +1687,8 @@ class CuZnInitialDatabase(InitialDatabase):
 
                         total_slabs_generated_count += len(gen_slabs)
                         total_slabs_generated.extend(gen_slabs)
-                        job_progress.update(sub_task, total=1)
-                        job_progress.advance(sub_task, advance=1)
+                        # job_progress.update(sub_task, total=1)
+                        # job_progress.advance(sub_task, advance=1)
 
                         overall_progress.advance(overall_task, advance=1)
 
@@ -2462,8 +1711,12 @@ class CuZnInitialDatabase(InitialDatabase):
 
                     # Generating a list of random percentages inside the current phase
                     # range.
-                    gen_percentages = self._gen_perc_surfaces(
-                        slab_phase, num_replacement_structs, current_perc, relative=True
+                    gen_percentages = mdb_surf.gen_perc_surfaces(
+                        phase=slab_phase,
+                        num_struct=num_replacement_structs,
+                        current_perc=current_perc,
+                        relative=True,
+                        db_obj=self,
                     )
 
                     # Going over the generated percentages
@@ -2491,7 +1744,7 @@ class CuZnInitialDatabase(InitialDatabase):
                                 supercell_vec_str_name = supercell_vec_str
 
                             # Creating a new Surface object for the structure with replacement
-                            new_struct_symm = mdf_struct.Surface(
+                            new_struct_symm = mdb_struct.Surface(
                                 material_name=f"{phase.prototype}_{phase.name}_surface-{supercell_vec_str_name}-{str_ind+1}_replacement-{repl + 1}",
                                 material_id=phase.prototype,
                                 surface_miller=supercell_vec_str,
@@ -2525,7 +1778,7 @@ class CuZnInitialDatabase(InitialDatabase):
                 f"Generated {total_slabs_generated_final_cnt} surfaces.", "done"
             )
 
-    def _get_main_elem_perc(self, phase: Phase, structure):
+    def _get_main_elem_perc(self, phase: mdb_pd.Phase, structure):
         """
         This function provides the percentage of the main element
         for a given structure
@@ -2558,12 +1811,12 @@ class CuZnInitialDatabase(InitialDatabase):
 
     def generate_surfaces_replacements(
         self,
-        phase: Phase,
+        phase: mdb_pd.Phase,
         num_struct: int,
         num_repeats: int,
     ):
         if isinstance(phase, str):
-            phase = CuZnInitialDatabase.CUZN_PHASES.get_phase(phase)
+            phase = CuZnInitialDatabase.DB_PHASE_DIAGRAM.get_phase(phase)
 
         # Instantiating RNG
         rng = np.random.default_rng()
@@ -2690,48 +1943,10 @@ class CuZnInitialDatabase(InitialDatabase):
         # TODO: Converting NaN values given by some functions to None
         # self.df.fillna(None, method=None, inplace=True)
 
-    # def gather_aiida_group_structures(self, group_name: str):
-    #     # Loading aiida profile
-    #     load_profile()
-
-    #     gathered_nodes = []
-
-    #     # Getting group
-    #     group =form.load_group(label=group_name)
-
-    #     # Storing every node contained in the group into a list
-    #     for node in group.nodes:
-    #         if isinstance(node,form.CalcJobNode):
-    #             gathered_nodes.append(node)
-    #         else:
-    #             for descendant in node.called_descendants:
-    #                 if isinstance(descendant,form.CalcJobNode):
-    #                     gathered_nodes.append(descendant)
-
-    #     return gathered_nodes
-
-    # def _get_pot_energy_outcar_aiida_node(self, vasprun:vasp.Vasprun) -> float:
-    #     # retrieved = node.outputs.retrieved
-    #     # vasprun = retrieved.get_object_content("vasp_run.xml")
-
-    #     # with open("/tmp/vasprun.tmp", "w") as f:
-    #     #     f.write(vasprun)
-
-    #     # vasprun = vasp.Vasprun("/tmp/vasprun.tmp")
-
-    #     # Getting energy in eV
-    #     # This energy is given in VASP as: 'free energy TOTEN'
-    #     # n2p2 uses this energy
-    #     energy = float(vasprun.ionic_steps[-1]['e_fr_energy'])
-
-    #     # Converting to Ha
-    #     energy *= self.eV2Eh
-
-    #     return energy
-
     def _gather_n2p2_reqdata_from_node(self, node):
         # Getting calculation name
         name = node.label + "_aiida-uuid_" + node.uuid
+
         # WARNING:
         # Getting potential energy
         # misc = node.outputs.misc.get_dict()
@@ -2895,93 +2110,77 @@ class CuZnInitialDatabase(InitialDatabase):
                 self._add_entry_to_n2p2_input(buffer=curr_f, data_dict=data_dict)
         ut.custom_print(f"All calculations saved in '{path}'.", "done")
 
+    def limit_structure_number_phases(
+        self,
+        structure_limit: int,
+        structure_types: list,
+        phases_to_use: list = None,
+    ):
+        # Getting which phases to use
+        if phases_to_use:
+            num_phases = len(phases_to_use)
+        else:
+            phases_to_use = self.DB_PHASE_DIAGRAM.phases
+            num_phases = len(self.DB_PHASE_DIAGRAM.phases)
 
-def gather_secrets():
-    """
-    Gather Materials project API key from a secret.json file.
+        # Maximum number of structures per phase
+        max_struct_phase = structure_limit // num_phases
 
-    Notes
-    -----
-        The json file should have the following structure:
+        # Getting the current phase as a Phase object
+        # If the phase is not found, omitting.
+        for curr_phase in phases_to_use:
+            try:
+                curr_phase = self.DB_PHASE_DIAGRAM.get_phase(curr_phase)
+            except mdb_exc.PhaseNotFound:
+                pass
 
-        {
-            "API_KEY": "XXXXXX"
-        }
+            # Getting the selected types to use.
+            # TODO: Fix this, only keeps one structure type.
+            structure_list = self.df.loc[self.df.phase == curr_phase]
+            for structure_type in structure_types:
+                structure_list = structure_list.loc[structure_list[structure_type]]
 
+            # If the number of structures for the selected types is larger than
+            # the maximum allowed, reduce the number by sampling a certain amount.
+            if structure_list.shape[0] > max_struct_phase:
+                # Getting the all but the original surface used for this phase,
+                # which should be maintained in the database.
+                changed_structures = structure_list.loc[
+                    structure_list.perturb
+                    | structure_list.supercell
+                    | structure_list.replacement
+                ]
 
-    Returns
-    -------
-    dict
-        object containing the api key
-    """
-    initial_db_path = pathlib.Path(__file__).parent
+                structure_list_base = structure_list.drop(changed_structures.index)
 
-    if pathlib.Path("secrets.json").exists():
-        with open("secrets.json", "r") as f:
-            secrets = js.load(f)
-
-    elif pathlib.Path(initial_db_path, "secrets.json").exists():
-        path = pathlib.Path(initial_db_path, "secrets.json")
-        with open(path, "r") as f:
-            secrets = js.load(f)
-
-    else:
-        raise FileNotFoundError(
-            f"'secrets.json' not found!\nPlease, add a 'secrets.json' file in the"
-            f" following directory: '{initial_db_path}'. "
-        )
-        secrets = None
-
-    return secrets
-
-
-def check_incorrect_ratios(df, curr_phase_diag):
-    for id, row in df.iterrows():
-        if not row.base and not row.material_name.endswith("_symm"):
-            strct = row.structure.get_sorted_structure()
-            name = row.material_name
-            phase = curr_phase_diag.get_phase(row.phase)
-            offset = phase.offset
-            tot_atoms = len(strct.species)
-            one_at_perc = 1 / tot_atoms
-
-            tot_cu = strct.species.count(Species("Cu")) + strct.species.count(
-                Element("Cu")
-            )
-            tot_zn = strct.species.count(Species("Zn")) + strct.species.count(
-                Element("Zn")
-            )
-
-            # Checking the total atom number
-            assert (
-                tot_cu + tot_zn == tot_atoms
-            ), f"""Total count does not match.
-            tot_cu: {tot_cu}, tot_zn: {tot_zn}, total: {tot_atoms}.
-            Species: {set(strct.species)}"""
-
-            perc = round(tot_zn / tot_atoms, 2)
-
-            offset_min = round(phase.base_elem_comp_min - offset, 2)
-            if offset_min < 0:
-                offset_min = 0
-
-            offset_max = round(phase.base_elem_comp_max + offset, 2)
-            if offset_max > 1:
-                offset_max = 1
-
-            # Checking if the current structure is between the phase ratio
-            # percentages.
-            if not (offset_min <= perc <= offset_max):
-                # If the structure could be fixed by adding or removing an atom
-                if (offset_min <= perc + one_at_perc <= offset_max) or (
-                    offset_min <= perc - one_at_perc <= offset_max
-                ):
-                    ut.custom_print(
-                        f"{name}: {perc:.2f} Zn outside of ({offset_min:.2f} - {offset_max:.2f}) range",
-                        "error",
+                if changed_structures.shape[0] >= max_struct_phase // 2:
+                    phase_structures = structure_list_base.iloc[
+                        0 : max_struct_phase // 2
+                    ]
+                    changed_structures_sample = changed_structures.sample(
+                        max_struct_phase // 2
                     )
                 else:
-                    pass
+                    offset = (max_struct_phase // 2) - changed_structures.shape[0]
+                    phase_structures = structure_list_base.iloc[
+                        0 : (max_struct_phase // 2) + offset
+                    ]
+                    changed_structures_sample = changed_structures
+
+                # These are the selected structures for the desired phase and type.
+                # We want to keep these structures in the original dataframe.
+                phase_structures = pd.concat(
+                    [phase_structures, changed_structures_sample]
+                )
+
+                # We remove all structures of the selected type and phase from the original
+                # dataframe, and add the selected ones.
+                orig_removed = self.df.loc[
+                    ~self.df.unique_id.isin(structure_list.unique_id)
+                ]
+
+                phase_result_df = pd.concat([orig_removed, phase_structures])
+                self.df = phase_result_df
 
 
 if __name__ == "__main__":
